refactor: renames mask file to index.ts
