+++ src/pages/promoters/SupportFundPage/CardSection/BillingInformationSection/index.tsx
@@ -4,7 +4,7 @@ import InputAutoComplete from "components/atomics/inputs/InputAutoComplete";
 import { Languages } from "types/enums/Languages";
 import { useLanguage } from "hooks/useLanguage";
 import { countryList } from "./countryList";
-import { mask } from "../../../../../lib/maskForTaxId/mask";
+import { mask } from "../../../../../lib/maskForTaxId";
 import * as S from "./styles";
 
 function BillingInformationSection(): JSX.Element {
